Moved the creating of mongo queries to own file.
